requirements.txt
discord.py
coloredlogs